backend/app/pyproject.toml
[tool.poetry]
name = "stockk"
version = "0.1.0"
description = ""
authors = ["Xuan Hien <hiennguyen9874@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-multipart = "^0.0.5"
email-validator = "^1.3.0"
requests = "^2.28.1"
celery = "^5.2.7"
passlib = "^1.7.4"
bcrypt = "^4.0.1"
tenacity = "^8.1.0"
pydantic = "^1.10.2"
gunicorn = "^20.1.0"
alembic = "^1.8.1"
asyncpg = "^0.27.0"
psycopg2-binary = "^2.9.5"
SQLAlchemy = "^1.4.44"
importlib-resources = "^5.10.0"
importlib-metadata = "^5.1.0"
fastapi = "^0.87.0"
aiofiles = "^22.1.0"
shortuuid = "^1.0.11"
pytz = "^2022.6"
click = "^8.1.3"
ulid-py = "^1.1.0"
watchdog = "^2.2.0"
aioredis = "^2.0.1"
httpx = "^0.23.1"
numpy = "^1.23.5"
more-itertools = "^9.0.0"
sentry-sdk = {extras = ["fastapi"], version = "^1.11.1"}
uvicorn = {extras = ["standard"], version = "^0.20.0"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
fastapi-pagination = {extras = ["sqlalchemy"], version = "^0.11.0"}
Authlib = "^1.2.0"
itsdangerous = "^2.1.2"
fastapi-mail = "^1.2.2"
redis = {extras = ["hiredis"], version = "^4.4.0"}
loguru = "^0.6.0"
influxdb-client = {extras = ["async"], version = "^1.35.0"}
pandas = "^1.5.2"
types-requests = "^2.28.11.7"
asgiref = "^3.6.0"

[tool.poetry.dev-dependencies]
black = "^22.8.0"
pylint = "^2.15.3"
flake8 = "^5.0.4"
isort = "^5.10.1"
mypy = "^0.982"
pre-commit = "^2.20.0"
flake8-blind-except = "^0.2.1"
flake8-bandit = "^4.1.1"
flake8-docstrings = "^1.6.0"
flake8-logging-format = "^0.8.1"
flake8-print = "^5.0.0"
autoflake = "^1.6.1"
sqlalchemy-stubs = "^0.4"
types-geoip2 = "^3.0.0"
types-pytz = "^2022.4.0"
types-redis = "^4.3.21"
types-aiofiles = "^22.1.0"
bandit = "^1.7.4"

[tool.poetry.group.dev.dependencies]
types-pyyaml = "^6.0.12.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
